--- packages/adapter-utilities/src/denormalize-flag-name/denormalize-flag-name.spec.ts
@@ -1,12 +1,0 @@
-import type { TFlagName } from '@flopflip/types';
-import { describe, expect, it } from 'vitest';
-import { denormalizeFlagName } from './denormalize-flag-name';
-
-describe('with camel case', () => {
-  it('should camel case to uppercased flag names', () => {
-    const input: TFlagName = 'aFlag';
-    const expected: string = 'a-flag';
-
-    expect(denormalizeFlagName(input)).toEqual(expected);
-  });
-});
