--- packages/adapter-utilities/src/denormalize-flag-name/index.ts
@@ -1,1 +1,0 @@
-export { denormalizeFlagName } from './denormalize-flag-name';
