+++ packages/adapter-utilities/src/expose-globally.ts
