--- packages/adapter-utilities/src/expose-globally/index.ts
@@ -1,1 +1,0 @@
-export { exposeGlobally } from './expose-globally';
