+++ packages/adapter-utilities/src/normalize-flag.ts
