--- packages/adapter-utilities/src/normalize-flag/index.ts
@@ -1,1 +1,0 @@
-export { normalizeFlag } from './normalize-flag';
