+++ packages/adapter-utilities/src/normalize-flags.ts
@@ -1,6 +1,6 @@
 import type { TFlag, TFlagVariation, TFlags } from '@flopflip/types';
 
-import { normalizeFlag as defaultNormalizeFlag } from '../normalize-flag/normalize-flag';
+import { normalizeFlag as defaultNormalizeFlag } from './normalize-flag';
 
 const normalizeFlags = (
   rawFlags: TFlags,
