--- packages/adapter-utilities/src/normalize-flags/index.ts
@@ -1,1 +1,0 @@
-export { normalizeFlags } from './normalize-flags';
