+++ packages/adapter-utilities/test/denormalize-flag-name.spec.js
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { denormalizeFlagName } from './denormalize-flag-name';
+import { denormalizeFlagName } from '../src/denormalize-flag-name';
 
 describe('with camel case', () => {
   it('should camel case to uppercase flag names', () => {
