+++ packages/adapter-utilities/test/expose-globally.spec.js
@@ -1,7 +1,6 @@
 import getGlobalThis from 'globalthis';
 import { expect, it } from 'vitest';
-
-import { exposeGlobally } from './expose-globally';
+import { exposeGlobally } from '../src/expose-globally';
 
 const testAdapter = {
   id: 'test',
